Переименованы классы виджетов
