+++ CRM/leader/leaderWidget.cpp
@@ -1,4 +1,4 @@
-// Seller source code for CRM application
+// Leader source code for CRM application
 // Copyright(C) 2022 Plnguins
 
 // This program is free software : you can redistribute it and / or modify
@@ -13,16 +13,20 @@
 
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.If not, see < https:  // www.gnu.org/licenses/>.
-#include "seller.h"
+#include "leaderWidget.h"
 
-seller::seller(QMainWindow *parent)
-    : parent(dynamic_cast<MainWindow *>(parent)), ui(new Ui::sellerUi) {
+leaderWidget::leaderWidget(QMainWindow *parent)
+    : parent(dynamic_cast<MainWindow *>(parent)), ui(new Ui::leaderUi) {
     ui->setupUi(this);
 
-    ui->Client->setIcon(QIcon(":/images/client-white.png"));
-    ui->Client->setIconSize({30, 30});
+    ui->Provider->setIcon(QIcon(":/images/vendor-white.png"));
+    ui->Provider->setIconSize({30, 30});
     ui->Deal->setIcon(QIcon(":/images/dollar.png"));
     ui->Deal->setIconSize({27, 27});
+    ui->Ads->setIcon(QIcon(":/images/ad-white.png"));
+    ui->Ads->setIconSize({27, 27});
+    ui->Employee->setIcon(QIcon(":/images/worker-white.png"));
+    ui->Employee->setIconSize({30, 30});
     ui->Stock->setIcon(QIcon(":/images/sklad-white.png"));
     ui->Stock->setIconSize({30, 30});
     ui->Support->setIcon(QIcon(":/images/t-white.png"));
@@ -36,7 +40,7 @@ seller::seller(QMainWindow *parent)
     ui->Edit->hide();
     ui->Add->hide();
     ui->Delete->hide();
-    ui->Title->hide();
+    ui->ProviderLabel->hide();
     ui->Send->hide();
     ui->Help->hide();
     ui->textEdit->hide();
@@ -47,25 +51,44 @@ seller::seller(QMainWindow *parent)
     ui->Icon->setPixmap(icon.scaled(ui->Icon->size(), Qt::KeepAspectRatio));
 }
 
-seller::~seller() { delete ui; }
+leaderWidget::~leaderWidget() { delete ui; }
 
-void seller::on_Client_clicked() {
+void leaderWidget::on_Logout_clicked() {
+    ui->Icon->show();
+    ui->Title->show();
+    ui->Company->show();
+    ui->Greeting->show();
+    ui->tableWidget->hide();
+    ui->RoundedBlue->hide();
+    ui->Update->hide();
+    ui->Edit->hide();
+    ui->Add->hide();
+    ui->Delete->hide();
+    ui->ProviderLabel->hide();
+    ui->Send->hide();
+    ui->Help->hide();
+    ui->textEdit->hide();
+    ui->textEdit->clear();
+    parent->setLoginInterface();
+}
+
+void leaderWidget::on_Provider_clicked() {
     ui->tableWidget->show();
     ui->Update->show();
     ui->Edit->show();
     ui->Add->show();
     ui->Delete->show();
     ui->RoundedBlue->show();
-    ui->Title->show();
+    ui->ProviderLabel->show();
     ui->Icon->hide();
-    ui->Name->hide();
+    ui->Title->hide();
     ui->Company->hide();
     ui->Greeting->hide();
     ui->Send->hide();
     ui->Help->hide();
     ui->textEdit->hide();
 
-    ui->Title->setText("Клиенты");
+    ui->ProviderLabel->setText("Поставщики");
     ui->Update->setText("Update");
 
     // TODO: fill table with db
@@ -73,21 +96,16 @@ void seller::on_Client_clicked() {
 
     // get from db
     ui->tableWidget->setRowCount(5);
-    ui->tableWidget->setColumnCount(7);
-    ui->tableWidget->setColumnWidth(0, 65);
-    ui->tableWidget->setColumnWidth(1, 96);
-    ui->tableWidget->setColumnWidth(2, 65);
-    ui->tableWidget->setColumnWidth(3, 65);
-    ui->tableWidget->setColumnWidth(4, 45);
-    ui->tableWidget->setColumnWidth(5, 65);
-    ui->tableWidget->setColumnWidth(6, 65);
-    QStringList Labels = {"id", "FIO", "Date", "City", "Sex", "e-mail", "tel"};
+    ui->tableWidget->setColumnCount(2);
+    ui->tableWidget->setColumnWidth(0, 91);
+    ui->tableWidget->setColumnWidth(1, 375);
+    QStringList Labels = {"id", "Name"};
     ui->tableWidget->setHorizontalHeaderLabels(Labels);
 
     // temporary
     QTableWidgetItem *item = new QTableWidgetItem("123");
     for (size_t i = 0; i < 5; i++) {
-        for (size_t j = 0; j < 7; j++) {
+        for (size_t j = 0; j < 2; j++) {
             QTableWidgetItem *item = new QTableWidgetItem("123");
             ui->tableWidget->setItem(i, j, item);
             ui->tableWidget->item(i, j)->setFlags(Qt::ItemIsEnabled |
@@ -95,29 +113,29 @@ void seller::on_Client_clicked() {
         }
     }
 
-    connect(ui->Update, SIGNAL(clicked()), this, SLOT(tableClientUpdate()));
+    connect(ui->Update, SIGNAL(clicked()), this, SLOT(tableVendorUpdate()));
     // connect(ui->Edit, SIGNAL(clicked()), this, SLOT(doSmth()));
     // connect(ui->Add, SIGNAL(clicked()), this, SLOT(doSmth()));
     // connect(ui->Delete, SIGNAL(clicked()), this, SLOT(doSmth()));
 }
 
-void seller::on_Deal_clicked() {
+void leaderWidget::on_Deal_clicked() {
     ui->tableWidget->show();
     ui->Update->show();
     ui->Edit->show();
     ui->Add->show();
     ui->Delete->show();
     ui->RoundedBlue->show();
-    ui->Title->show();
+    ui->ProviderLabel->show();
     ui->Icon->hide();
-    ui->Name->hide();
+    ui->Title->hide();
     ui->Company->hide();
     ui->Greeting->hide();
     ui->Send->hide();
     ui->Help->hide();
     ui->textEdit->hide();
 
-    ui->Title->setText("Сделки");
+    ui->ProviderLabel->setText("Сделки");
     ui->Update->setText("Update");
 
     // TODO: fill table with db
@@ -157,23 +175,23 @@ void seller::on_Deal_clicked() {
     // connect(ui->Delete, SIGNAL(clicked()), this, SLOT(doSmth()));
 }
 
-void seller::on_Stock_clicked() {
+void leaderWidget::on_Stock_clicked() {
     ui->tableWidget->show();
     ui->Update->show();
     ui->Edit->show();
     ui->Add->show();
     ui->Delete->show();
     ui->RoundedBlue->show();
-    ui->Title->show();
+    ui->ProviderLabel->show();
     ui->Icon->hide();
-    ui->Name->hide();
+    ui->Title->hide();
     ui->Company->hide();
     ui->Greeting->hide();
     ui->Send->hide();
     ui->Help->hide();
     ui->textEdit->hide();
 
-    ui->Title->setText("Склад");
+    ui->ProviderLabel->setText("Склад");
     ui->Update->setText("Update");
 
     // TODO: fill table with db
@@ -210,49 +228,141 @@ void seller::on_Stock_clicked() {
     // connect(ui->Delete, SIGNAL(clicked()), this, SLOT(doSmth()));
 }
 
-void seller::on_Support_clicked() {
+void leaderWidget::on_Ads_clicked() {
+    ui->tableWidget->show();
+    ui->Update->show();
+    ui->Edit->show();
+    ui->Add->show();
+    ui->Delete->show();
+    ui->RoundedBlue->show();
+    ui->ProviderLabel->show();
+    ui->Icon->hide();
+    ui->Title->hide();
+    ui->Company->hide();
+    ui->Greeting->hide();
+    ui->Send->hide();
+    ui->Help->hide();
+    ui->textEdit->hide();
+
+    ui->ProviderLabel->setText("Реклама");
+    ui->Update->setText("Update");
+
+    // TODO: fill table with db
+    // tableVendorUpdate();
+
+    // get from db
+    ui->tableWidget->setRowCount(5);
+    ui->tableWidget->setColumnCount(2);
+    ui->tableWidget->setColumnWidth(0, 91);
+    ui->tableWidget->setColumnWidth(1, 375);
+    QStringList Labels = {"id", "Place"};
+    ui->tableWidget->setHorizontalHeaderLabels(Labels);
+
+    // temporary
+    QTableWidgetItem *item = new QTableWidgetItem("123");
+    for (size_t i = 0; i < 5; i++) {
+        for (size_t j = 0; j < 2; j++) {
+            QTableWidgetItem *item = new QTableWidgetItem("123");
+            ui->tableWidget->setItem(i, j, item);
+            ui->tableWidget->item(i, j)->setFlags(Qt::ItemIsEnabled |
+                                                  Qt::ItemIsSelectable);
+        }
+    }
+
+    connect(ui->Update, SIGNAL(clicked()), this, SLOT(tableAdUpdate()));
+    // connect(ui->Edit, SIGNAL(clicked()), this, SLOT(doSmth()));
+    // connect(ui->Add, SIGNAL(clicked()), this, SLOT(doSmth()));
+    // connect(ui->Delete, SIGNAL(clicked()), this, SLOT(doSmth()));
+}
+
+void leaderWidget::on_Employee_clicked() {
+    ui->tableWidget->show();
+    ui->Update->show();
+    ui->Edit->show();
+    ui->Add->show();
+    ui->Delete->show();
+    ui->RoundedBlue->show();
+    ui->ProviderLabel->show();
+    ui->Icon->hide();
+    ui->Title->hide();
+    ui->Company->hide();
+    ui->Greeting->hide();
+    ui->Send->hide();
+    ui->Help->hide();
+    ui->textEdit->hide();
+
+    ui->ProviderLabel->setText("Работники");
+    ui->Update->setText("Update");
+
+    // TODO: fill table with db
+    // tableVendorUpdate();
+
+    // get from db
+    ui->tableWidget->setRowCount(5);
+    ui->tableWidget->setColumnCount(5);
+    ui->tableWidget->setColumnWidth(0, 90);
+    ui->tableWidget->setColumnWidth(1, 136);
+    ui->tableWidget->setColumnWidth(2, 90);
+    ui->tableWidget->setColumnWidth(3, 90);
+    ui->tableWidget->setColumnWidth(4, 60);
+    QStringList Labels = {"id", "FIO", "Login", "Password", "Number"};
+    ui->tableWidget->setHorizontalHeaderLabels(Labels);
+
+    // temporary
+    QTableWidgetItem *item = new QTableWidgetItem("123");
+    for (size_t i = 0; i < 5; i++) {
+        for (size_t j = 0; j < 5; j++) {
+            QTableWidgetItem *item = new QTableWidgetItem("123");
+            ui->tableWidget->setItem(i, j, item);
+            ui->tableWidget->item(i, j)->setFlags(Qt::ItemIsEnabled |
+                                                  Qt::ItemIsSelectable);
+        }
+    }
+
+    connect(ui->Update, SIGNAL(clicked()), this, SLOT(tableVendorUpdate()));
+    // connect(ui->Edit, SIGNAL(clicked()), this, SLOT(doSmth()));
+    // connect(ui->Add, SIGNAL(clicked()), this, SLOT(doSmth()));
+    // connect(ui->Delete, SIGNAL(clicked()), this, SLOT(doSmth()));
+}
+
+void leaderWidget::on_Support_clicked() {
     ui->tableWidget->hide();
     ui->Update->hide();
     ui->Edit->hide();
     ui->Add->hide();
     ui->Delete->hide();
     ui->RoundedBlue->hide();
-    ui->Title->show();
+    ui->ProviderLabel->show();
     ui->Icon->hide();
-    ui->Name->hide();
+    ui->Title->hide();
     ui->Company->hide();
     ui->Greeting->hide();
     ui->Send->show();
     ui->Help->show();
     ui->textEdit->show();
 
-    ui->Title->setText("ТехПод");
+    ui->ProviderLabel->setText("ТехПод");
     ui->Send->setText("Отправить");
 }
 
-void seller::on_Logout_clicked() {
-    ui->Icon->show();
-    ui->Name->show();
-    ui->Company->show();
-    ui->Greeting->show();
-    ui->tableWidget->hide();
-    ui->RoundedBlue->hide();
-    ui->Update->hide();
-    ui->Edit->hide();
-    ui->Add->hide();
-    ui->Delete->hide();
-    ui->Title->hide();
-    ui->Send->hide();
-    ui->Help->hide();
-    ui->textEdit->hide();
-    ui->textEdit->clear();
-    parent->setLoginInterface();
-}
+//----------table update functions---------------
 
-void seller::tableClientUpdate() {
+void leaderWidget::tableVendorUpdate() {
     //
 }
 
-void seller::tableStorageUpdate() {}
+void leaderWidget::tableDealUpdate() {
+    //
+}
 
-void seller::tableDealUpdate() {}
+void leaderWidget::tableStorageUpdate() {
+    //
+}
+
+void leaderWidget::tableAdUpdate() {
+    //
+}
+
+void leaderWidget::tableEmployeesUpdate() {
+    //
+}
