+++ CRM/seller/sellerWidget.h
@@ -1,4 +1,4 @@
-// Manager window header for CRM application
+// Seller window header for CRM application
 // Copyright(C) 2022 Plnguins
 
 // This program is free software : you can redistribute it and / or modify
@@ -19,30 +19,24 @@
 #include <QWidget>
 
 #include "../mainwindow.h"
-#include "ui_manager.h"
+#include "ui_seller.h"
 
 namespace Ui {
-class managerUi;
+class sellerUi;
 }
 
-class manager : public QWidget {
+class sellerWidget : public QWidget {
     Q_OBJECT
 
    public:
-    explicit manager(QMainWindow *parent = nullptr);
-    ~manager();
+    explicit sellerWidget(QMainWindow *parent = nullptr);
+    ~sellerWidget();
 
    signals:
     void logout();
 
    private slots:
-    void tableVendorUpdate();
-    void tableDealUpdate();
-    void tableStorageUpdate();
-
-    void on_Logout_clicked();
-
-    void on_Provider_clicked();
+    void on_Client_clicked();
 
     void on_Deal_clicked();
 
@@ -50,7 +44,13 @@ class manager : public QWidget {
 
     void on_Support_clicked();
 
+    void on_Logout_clicked();
+
+    void tableClientUpdate();
+    void tableStorageUpdate();
+    void tableDealUpdate();
+
    private:
     MainWindow *parent;
-    Ui::managerUi *ui;
+    Ui::sellerUi *ui;
 };
